put class using LibFlute in its own file
